insert redirect into delete.php

+++ includes/delete.php
@@ -1,10 +1,16 @@
 <?php
-mysql_connect("localhost","root","california");
-mysql_select_db("print3D");
-if(isset($_GET['delete_id']))
+if(!$_SESSION['email'])
 {
- $sql_query="DELETE FROM prenotazioni WHERE id=".$_GET['delete_id'];
- mysql_query($sql_query);
- header("Location: ../php/prenotazioni.php");
+    header('Location: ../index.php');
+}
+else{
+ mysql_connect("localhost","root","california");
+ mysql_select_db("print3D");
+    if(isset($_GET['delete_id']))
+    {
+     $sql_query="DELETE FROM prenotazioni WHERE id=".$_GET['delete_id'];
+     mysql_query($sql_query);
+     header("Location: ../php/prenotazioni.php");
+     }
 }
 ?>
